Move governance query commands to query command group.
