+++ cardano-cli/test/cardano-cli-golden/files/golden/help/conway_governance.cli
@@ -1,4 +1,4 @@
-Usage: cardano-cli conway governance (query | action | committee | drep | vote)
+Usage: cardano-cli conway governance (action | committee | drep | vote)
 
   Governance commands.
 
@@ -6,7 +6,6 @@ Available options:
   -h,--help                Show this help text
 
 Available commands:
-  query                    Query governance-related information
   action                   Governance action commands.
   committee                Committee member commands.
   drep                     DRep member commands.
