--- cardano-cli/test/cardano-cli-golden/files/golden/help/conway_governance_query.cli
@@ -1,22 +1,0 @@
-Usage: cardano-cli conway governance query 
-                                             ( constitution
-                                             | gov-state
-                                             | drep-state
-                                             | drep-stake-distribution
-                                             | committee-state
-                                             )
-
-  Query governance-related information
-
-Available options:
-  -h,--help                Show this help text
-
-Available commands:
-  constitution             Get the constitution
-  gov-state                Get the governance state
-  drep-state               Get the DRep state. If no DRep credentials are
-                           provided, return states for all of them.
-  drep-stake-distribution  Get the DRep stake distribution. If no DRep
-                           credentials are provided, return stake distributions
-                           for all of them.
-  committee-state          Get the committee state
