+++ cardano-cli/test/cardano-cli-golden/files/golden/help/conway_query_committee-state.cli
@@ -1,15 +1,14 @@
-Usage: cardano-cli conway governance query committee-state 
-                                                             --socket-path SOCKET_PATH
-                                                             [ --shelley-mode
-                                                             | --byron-mode
-                                                               [--epoch-slots SLOTS]
-                                                             | --cardano-mode
-                                                               [--epoch-slots SLOTS]
-                                                             ]
-                                                             ( --mainnet
-                                                             | --testnet-magic NATURAL
-                                                             )
-                                                             [--out-file FILE]
+Usage: cardano-cli conway query committee-state --socket-path SOCKET_PATH
+                                                  [ --shelley-mode
+                                                  | --byron-mode
+                                                    [--epoch-slots SLOTS]
+                                                  | --cardano-mode
+                                                    [--epoch-slots SLOTS]
+                                                  ]
+                                                  ( --mainnet
+                                                  | --testnet-magic NATURAL
+                                                  )
+                                                  [--out-file FILE]
 
   Get the committee state
