+++ cardano-cli/test/cardano-cli-golden/files/golden/help/conway_query_drep-state.cli
@@ -1,22 +1,21 @@
-Usage: cardano-cli conway governance query drep-stake-distribution 
-                                                                     --socket-path SOCKET_PATH
-                                                                     [ --shelley-mode
-                                                                     | --byron-mode
-                                                                       [--epoch-slots SLOTS]
-                                                                     | --cardano-mode
-                                                                       [--epoch-slots SLOTS]
-                                                                     ]
-                                                                     ( --mainnet
-                                                                     | --testnet-magic NATURAL
-                                                                     )
-                                                                     ( --drep-verification-key STRING
-                                                                     | --drep-verification-key-file FILE
-                                                                     | --drep-key-hash HASH
-                                                                     )
-                                                                     [--out-file FILE]
+Usage: cardano-cli conway query drep-state --socket-path SOCKET_PATH
+                                             [ --shelley-mode
+                                             | --byron-mode
+                                               [--epoch-slots SLOTS]
+                                             | --cardano-mode
+                                               [--epoch-slots SLOTS]
+                                             ]
+                                             ( --mainnet
+                                             | --testnet-magic NATURAL
+                                             )
+                                             ( --drep-verification-key STRING
+                                             | --drep-verification-key-file FILE
+                                             | --drep-key-hash HASH
+                                             )
+                                             [--out-file FILE]
 
-  Get the DRep stake distribution. If no DRep credentials are provided, return
-  stake distributions for all of them.
+  Get the DRep state. If no DRep credentials are provided, return states for all
+  of them.
 
 Available options:
   --socket-path SOCKET_PATH
