+++ cardano-cli/test/cardano-cli-golden/files/golden/help/conway_query_gov-state.cli
@@ -1,14 +1,13 @@
-Usage: cardano-cli conway governance query gov-state --socket-path SOCKET_PATH
-                                                       [ --shelley-mode
-                                                       | --byron-mode
-                                                         [--epoch-slots SLOTS]
-                                                       | --cardano-mode
-                                                         [--epoch-slots SLOTS]
-                                                       ]
-                                                       ( --mainnet
-                                                       | --testnet-magic NATURAL
-                                                       )
-                                                       [--out-file FILE]
+Usage: cardano-cli conway query gov-state --socket-path SOCKET_PATH
+                                            [ --shelley-mode
+                                            | --byron-mode [--epoch-slots SLOTS]
+                                            | --cardano-mode
+                                              [--epoch-slots SLOTS]
+                                            ]
+                                            ( --mainnet
+                                            | --testnet-magic NATURAL
+                                            )
+                                            [--out-file FILE]
 
   Get the governance state
